Changes to base model: FPOs instead of SFPs
Stopped considering the case of Single False Positives (SFPs): it was
difficult to calculate the probability of a single FP + the case of
having only FP matches (either 1 or multiple) covers the intended
case.

Notice that if presented with multiple FP matches, a router will only
pick one (e.g. the longest prefix), so it doesn't really matter if
it is just 1 match or a bunch of them.